Refactor : widget 컴포넌트 features로 변경
