+++ src/features/comments/add-comment/index.ts
@@ -1,0 +1,2 @@
+export * from "./ui/AddCommentForm"
+export * from "./model/useAddComment"
