+++ src/features/comments/edit-comment/index.ts
@@ -1,0 +1,2 @@
+export * from "./ui/EditCommentForm"
+export * from "./model/useEditComment"
