+++ src/features/posts/add-post/index.ts
@@ -1,0 +1,2 @@
+export * from "./ui/AddPostForm"
+export * from "./model/useAddPost"
