+++ src/features/posts/edit-post/index.ts
@@ -1,0 +1,2 @@
+export * from "./ui/EditPostForm"
+export * from "./model/useEditPost"
