+++ src/features/posts/post-detail/index.ts
@@ -1,0 +1,2 @@
+export * from "./ui/PostDetailDialog"
+export * from "./model/usePostDetail"
