+++ src/features/posts/post-detail/model/usePostDetail.ts
@@ -1,0 +1,1 @@
+//게시물 상세 로직
