+++ src/features/posts/post-detail/ui/PostDetailDialog.tsx
@@ -1,0 +1,30 @@
+import { Dialog, DialogContent, DialogTitle } from "../../../../shared/ui"
+import { ReactNode } from "react"
+import { Post } from "../../../../entities/posts"
+import { useSearchFilter, usePostDialogs } from "../../../../shared/store"
+import { DialogHeader, highlightText } from "../../../../shared/ui"
+
+interface PostDetailDialogProps {
+  selectedPost: Post
+  children: ReactNode
+}
+
+export const PostDetailDialog = (props: PostDetailDialogProps) => {
+  const { selectedPost, children } = props
+  const { searchQuery } = useSearchFilter()
+  const { showPostDetailDialog, setShowPostDetailDialog } = usePostDialogs()
+
+  return (
+    <Dialog open={showPostDetailDialog} onOpenChange={setShowPostDetailDialog}>
+      <DialogContent className="max-w-3xl">
+        <DialogHeader>
+          <DialogTitle>{highlightText(selectedPost?.title, searchQuery)}</DialogTitle>
+        </DialogHeader>
+        <div className="space-y-4">
+          <p>{highlightText(selectedPost?.body, searchQuery)}</p>
+          {children}
+        </div>
+      </DialogContent>
+    </Dialog>
+  )
+}
