+++ src/features/posts/post-list/index.ts
@@ -1,0 +1,2 @@
+export * from './ui/PostList'
+export * from './model/usePostList'
