+++ src/features/posts/post-list/model/usePostList.ts
@@ -1,0 +1,1 @@
+// 추후 게시물 목록 관리 로직이 들어갈 예정
