+++ src/features/posts/post-list/ui/PostList.tsx
@@ -1,0 +1,115 @@
+import { Edit2, MessageSquare, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react"
+import {
+  TableHeader,
+  Table,
+  TableRow,
+  TableHead,
+  TableBody,
+  TableCell,
+  Button,
+  highlightText,
+} from "../../../../shared/ui"
+import { useSearchFilter } from "../../../../shared/store"
+import { Post } from "../../../../entities/posts/api/types"
+import { Users } from "../../../../entities/users/api/types"
+
+export interface PostsWithUsers extends Post {
+  author?: Users | undefined
+}
+
+interface PostListProps {
+  posts: Array<PostsWithUsers>
+  selectedTag: string
+  openUserModal: (userId: number) => Promise<void>
+  openPostDetail: (post: PostsWithUsers) => void
+  deletePost: (postId: number) => Promise<void>
+  filteredPostTag: (tagName: string) => void
+  openEditDialog: (selectedPost: PostsWithUsers, isEditDialog: boolean) => void
+}
+
+export const PostList = (props: PostListProps) => {
+  const { posts, selectedTag, openUserModal, openPostDetail, filteredPostTag, deletePost, openEditDialog } = props
+
+  const { searchQuery } = useSearchFilter()
+
+  return (
+    <Table>
+      <TableHeader>
+        <TableRow>
+          <TableHead className="w-[50px]">ID</TableHead>
+          <TableHead>제목</TableHead>
+          <TableHead className="w-[150px]">작성자</TableHead>
+          <TableHead className="w-[150px]">반응</TableHead>
+          <TableHead className="w-[150px]">작업</TableHead>
+        </TableRow>
+      </TableHeader>
+      <TableBody>
+        {posts.map((post) => (
+          <TableRow key={post.id}>
+            <TableCell>{post.id}</TableCell>
+            <TableCell>
+              <div className="space-y-1">
+                <div>{highlightText(post.title, searchQuery)}</div>
+
+                <div className="flex flex-wrap gap-1">
+                  {post.tags?.map((tag) => (
+                    <span
+                      key={tag}
+                      className={`px-1 text-[9px] font-semibold rounded-[4px] cursor-pointer ${
+                        selectedTag === tag
+                          ? "text-white bg-blue-500 hover:bg-blue-600"
+                          : "text-blue-800 bg-blue-100 hover:bg-blue-200"
+                      }`}
+                      onClick={() => filteredPostTag(tag)}
+                    >
+                      {tag}
+                    </span>
+                  ))}
+                </div>
+              </div>
+            </TableCell>
+            <TableCell>
+              <div
+                className="flex items-center space-x-2 cursor-pointer"
+                onClick={() => {
+                  if (!post.author) return
+                  openUserModal(post.author.id)
+                }}
+              >
+                <img src={post.author?.image} alt={post.author?.username} className="w-8 h-8 rounded-full" />
+                <span>{post.author?.username}</span>
+              </div>
+            </TableCell>
+            <TableCell>
+              <div className="flex items-center gap-2">
+                <ThumbsUp className="w-4 h-4" />
+                <span>{post.reactions?.likes || 0}</span>
+                <ThumbsDown className="w-4 h-4" />
+                <span>{post.reactions?.dislikes || 0}</span>
+              </div>
+            </TableCell>
+            <TableCell>
+              <div className="flex items-center gap-2">
+                <Button variant="ghost" size="sm" onClick={() => openPostDetail(post)}>
+                  <MessageSquare className="w-4 h-4" />
+                </Button>
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  onClick={() => {
+                    openEditDialog(post, true)
+                  }}
+                >
+                  <Edit2 className="w-4 h-4" />
+                </Button>
+                <Button variant="ghost" size="sm" onClick={() => deletePost(post.id)}>
+                  <Trash2 className="w-4 h-4" />
+                </Button>
+              </div>
+            </TableCell>
+          </TableRow>
+        ))}
+      </TableBody>
+    </Table>
+  )
+}
