+++ src/features/user/user-info/index.ts
@@ -1,0 +1,2 @@
+export * from "./ui/UserInfoDialog"
+export * from "./model/useUserInfo"
