--- src/widgets/add-comment-dialog/ui/AddCommentDialog.tsx
@@ -1,35 +1,0 @@
-import { Dispatch, SetStateAction } from "react"
-import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Textarea } from "../../../shared/ui"
-import { useCommentDialogs } from "../../../shared/store"
-import { CreateCommentsRequest } from "../../../entities/comments/model/types"
-
-export type NewComment = CreateCommentsRequest
-
-interface AddCommentDialogProps {
-  newComment: NewComment
-  setNewComment: Dispatch<SetStateAction<NewComment>>
-  addComment: () => Promise<void>
-}
-
-export const AddCommentDialog = (props: AddCommentDialogProps) => {
-  const { newComment, setNewComment, addComment } = props
-  const { showAddCommentDialog, setShowAddCommentDialog } = useCommentDialogs()
-
-  return (
-    <Dialog open={showAddCommentDialog} onOpenChange={setShowAddCommentDialog}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>새 댓글 추가</DialogTitle>
-        </DialogHeader>
-        <div className="space-y-4">
-          <Textarea
-            placeholder="댓글 내용"
-            value={newComment.body}
-            onChange={(e) => setNewComment({ ...newComment, body: e.target.value })}
-          />
-          <Button onClick={addComment}>댓글 추가</Button>
-        </div>
-      </DialogContent>
-    </Dialog>
-  )
-}
