+++ src/widgets/index.ts
@@ -1,11 +1,3 @@
 export * from "./layout/Footer"
 export * from "./layout/Header"
-export * from "./post-table/ui/PostTable"
-export * from "./add-post-dialog/ui/AddPostDialog"
-export * from "./edit-post-dialog/ui/EditPostDialog"
-export * from "./add-comment-dialog/ui/AddCommentDialog"
-export * from "./edit-comment-dialog/ui/EditCommentDialog"
 export * from "./pagination/ui/Pagination"
-export * from "./post-comments/ui/PostComments"
-export * from "./post-detail-dialog/ui/PostDetailDialog"
-export * from "./user-dialog/ui/UserDialog"
